Multiply Strings - big int support

+++ src/main/kotlin/MultiplyStrings.kt
@@ -1,5 +1,5 @@
 class MultiplyStrings {
     fun multiply(num1: String, num2: String): String {
-        return (num1.toInt() * num2.toInt()).toString()
+        return (num1.toBigInteger() * num2.toBigInteger()).toString()
     }
 }
